+++ src/metatrain/pet/trainer.py
@@ -18,9 +18,8 @@ from metatrain.utils.data import (
     Dataset,
     build_train_dataloaders,
     build_val_dataloaders,
-    get_num_workers,
+    resolve_dataloader_workers,
     unpack_batch,
-    validate_num_workers,
 )
 from metatrain.utils.data.atomic_basis_helpers import (
     get_prepare_atomic_basis_targets_transform,
@@ -300,17 +299,11 @@ class Trainer(TrainerInterface[TrainerHypers]):
             ],
         )
 
-        if self.hypers["num_workers"] is None:
-            num_workers = get_num_workers()
-            logging.info(
-                "Number of workers for data-loading not provided and chosen "
-                f"automatically. Using {num_workers} workers."
-            )
-        else:
-            num_workers = self.hypers["num_workers"]
-            validate_num_workers(num_workers)
-
         # Create dataloader for the training datasets:
+        num_workers, mp_context = resolve_dataloader_workers(
+            self.hypers["num_workers"], device, train_datasets + val_datasets
+        )
+
         train_dataloaders, epoch_samplers = build_train_dataloaders(
             train_datasets=train_datasets,
             train_distributed_samplers=train_samplers,
@@ -319,6 +312,7 @@ class Trainer(TrainerInterface[TrainerHypers]):
             max_atoms_per_batch=max_atoms,
             min_atoms_per_batch=self.hypers["min_atoms_per_batch"],
             num_workers=num_workers,
+            multiprocessing_context=mp_context,
         )
         train_dataloader = CombinedDataLoader(train_dataloaders, shuffle=True)
 
@@ -330,6 +324,7 @@ class Trainer(TrainerInterface[TrainerHypers]):
             batch_size=self.hypers["batch_size"],
             max_atoms_per_batch=max_atoms,
             num_workers=num_workers,
+            multiprocessing_context=mp_context,
         )
         val_dataloader = CombinedDataLoader(val_dataloaders, shuffle=False)